Prueba hosting

+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/Security/Service/UsuarioService.java
@@ -29,7 +29,7 @@ public class UsuarioService {
     }
     
    
-    public void deleteUsuario(Long id){
+    public void deleteUsuario(Integer id){
         usuarioRepository.deleteById(id);
     }
